Move all wpt code into //src/wpt
