--- src/workerd/api/wpt/eslint.config.mjs
@@ -1,11 +1,0 @@
-import { baseConfig } from '../../../../tools/base.eslint.config.mjs';
-
-export default [
-  ...baseConfig({ tsconfigRootDir: import.meta.dirname }),
-  {
-    files: ['src/workerd/api/wpt/**/*-test.ts'],
-    rules: {
-      'sort-keys': 'error',
-    },
-  },
-];
