+++ src/wpt/eslint.config.mjs
@@ -1,3 +1,11 @@
 import { baseConfig } from '../../tools/base.eslint.config.mjs';
 
-export default [...baseConfig({ tsconfigRootDir: import.meta.dirname })];
+export default [
+  ...baseConfig({ tsconfigRootDir: import.meta.dirname }),
+  {
+    files: ['src/wpt/**/*-test.ts'],
+    rules: {
+      'sort-keys': 'error',
+    },
+  },
+];
